fix: harden deps UTF-8 and split registries
