+++ dongle-smartcontract/src/review_registry/storage.rs
@@ -1,33 +1,26 @@
-//! Review registry: create/update/delete reviews and maintain aggregates and indexes.
+//! Review registry storage mutations: CRUD, moderation, aggregates, and listing.
 
 use crate::admin_action_log::AdminActionLog;
 use crate::constants::{
-    MAX_CID_LEN, MAX_PAGE_LIMIT, MAX_REVIEWS_PER_PROJECT, MAX_REVIEWS_PER_USER,
-    MAX_REVIEW_REVISIONS, RATING_MAX, RATING_MIN, REVIEW_UPDATE_COOLDOWN_SECONDS,
+    MAX_PAGE_LIMIT, MAX_REVIEWS_PER_PROJECT, MAX_REVIEWS_PER_USER, MAX_REVIEW_REVISIONS,
+    REVIEW_UPDATE_COOLDOWN_SECONDS,
 };
 use crate::errors::ContractError;
 use crate::events::{publish_review_event, publish_review_revision_event};
 use crate::project_registry::ProjectRegistry;
 use crate::rating_calculator::RatingCalculator;
+use crate::review_registry::validation::ReviewValidation;
 use crate::storage_keys::{ExtensionKey, StorageKey};
 use crate::storage_manager::StorageManager;
 use crate::types::{
-    AdminActionType, Project, ProjectStats, Review, ReviewAction, ReviewRevision, ReviewSortMode,
+    AdminActionType, ProjectStats, Review, ReviewAction, ReviewRevision, ReviewSortMode,
     ReviewTombstone,
 };
-use crate::utils::Utils;
 use soroban_sdk::{Address, Env, String, Vec};
 
 pub struct ReviewRegistry;
 
 impl ReviewRegistry {
-    fn validate_review_cid(cid: &String) -> Result<(), ContractError> {
-        if !Utils::is_valid_ipfs_cid(cid) || cid.len() as usize > MAX_CID_LEN {
-            return Err(ContractError::InvalidProjectData);
-        }
-        Ok(())
-    }
-
     pub fn add_review(
         env: &Env,
         project_id: u64,
@@ -36,7 +29,7 @@ impl ReviewRegistry {
         comment_cid: Option<String>,
     ) -> Result<(), ContractError> {
         if let Some(cid) = comment_cid.as_ref() {
-            Self::validate_review_cid(cid)?;
+            ReviewValidation::validate_review_cid(cid)?;
         }
 
         // Validation phase
@@ -49,18 +42,14 @@ impl ReviewRegistry {
         };
 
         // Project owners cannot review their own project
-        if project.owner == reviewer {
-            return Err(ContractError::OwnerCannotReview);
-        }
+        ReviewValidation::ensure_not_owner(&project, &reviewer)?;
 
         // Check if reviews are enabled for this project
         if !Self::get_reviews_enabled(env, project_id) {
             return Err(ContractError::ReviewsDisabled);
         }
 
-        if !(RATING_MIN..=RATING_MAX).contains(&rating) {
-            return Err(ContractError::InvalidRating);
-        }
+        ReviewValidation::validate_rating(rating)?;
 
         let review_key = StorageKey::Review(project_id, reviewer.clone());
         if env.storage().persistent().has(&review_key) {
@@ -164,7 +153,7 @@ impl ReviewRegistry {
         rating: u32,
         review_cid: String,
     ) -> Result<(), ContractError> {
-        Self::validate_review_cid(&review_cid)?;
+        ReviewValidation::validate_review_cid(&review_cid)?;
         Self::add_review(env, project_id, reviewer, rating, Some(review_cid))
     }
 
@@ -176,7 +165,7 @@ impl ReviewRegistry {
         comment_cid: Option<String>,
     ) -> Result<(), ContractError> {
         if let Some(cid) = comment_cid.as_ref() {
-            Self::validate_review_cid(cid)?;
+            ReviewValidation::validate_review_cid(cid)?;
         }
 
         // Validation phase
@@ -187,9 +176,7 @@ impl ReviewRegistry {
             return Err(ContractError::ProjectNotFound);
         }
 
-        if !(RATING_MIN..=RATING_MAX).contains(&rating) {
-            return Err(ContractError::InvalidRating);
-        }
+        ReviewValidation::validate_rating(rating)?;
 
         let review_key = StorageKey::Review(project_id, reviewer.clone());
         let mut review: Review = env
